+++ source/TestCommon/source/FunctionApp.TestCommon.Tests/Integration/Azurite/AzuriteManagerTests.cs
@@ -171,7 +171,7 @@ public class AzuriteManagerTests
             // Arrange
             var client = new BlobServiceClient(
                 serviceUri: Fixture.AzuriteManager!.BlobStorageServiceUri,
-                credential: new DefaultAzureCredential(),
+                credential: Fixture.Credential,
                 CreateBlobNoRetryOptions());
 
             // Act
@@ -217,7 +217,7 @@ public class AzuriteManagerTests
             // Arrange
             var client = new QueueServiceClient(
                 serviceUri: Fixture.AzuriteManager!.QueueStorageServiceUri,
-                credential: new DefaultAzureCredential(),
+                credential: Fixture.Credential,
                 CreateQueueNoRetryOptions());
 
             // Act
@@ -263,7 +263,7 @@ public class AzuriteManagerTests
             // Arrange
             var client = new TableServiceClient(
                 endpoint: Fixture.AzuriteManager!.TableStorageServiceUri,
-                tokenCredential: new DefaultAzureCredential(),
+                tokenCredential: Fixture.Credential,
                 CreateTableNoRetryOptions());
 
             // Act